add tinymce

+++ src/pages/Home.js
@@ -1,8 +1,8 @@
 import { React, useState, useEffect } from 'react'
-import { collection, doc, deleteDoc, query, startAfter, orderBy, limit } from "firebase/firestore";
+import { collection, doc, deleteDoc, query, startAfter, orderBy, limit, where, updateDoc } from "firebase/firestore";
 import { auth, db } from '../firebase-config';
-import { Button, Center, Box, Text, Flex, Spacer, Heading  } from '@chakra-ui/react'
-import { CloseIcon, EditIcon  } from '@chakra-ui/icons'
+import { Button, Center, Box, Text, Flex, Spacer, Heading, Input  } from '@chakra-ui/react'
+import { CloseIcon, EditIcon, ArrowUpIcon, DeleteIcon } from '@chakra-ui/icons'
 import { useNavigate } from "react-router-dom";
 import { signOut } from "firebase/auth";
 import { useScrollPosition } from '../hooks/useScrollDirection'
@@ -10,17 +10,43 @@ import ABasePage from './ABasePage'
 import { useFirestoreInfiniteQuery } from "@react-query-firebase/firestore";
 import { useSelector, useDispatch } from 'react-redux'
 import { save } from '../reducer/scrollReducer';
+import { setKey } from '../reducer/searchReducer';
+import { openTrash } from '../reducer/trashReducer';
+import { async } from '@firebase/util';
 
 
 const ListPost = ({isAuth}) => {
     const navigate = useNavigate()
     const scroll = useSelector((state) => state.scroll.value);
+    const search = useSelector((state) => state.search.value);
+    const trash = useSelector((state) => state.trash.value);
     const dispatch = useDispatch()
 
+    const [searchKey, setSearchKey] = useState(search.searchKey)
+
     const collectionRef = collection(db, 'posts')
-    const postsQuery = query(collectionRef, orderBy("createTime", "desc"), limit(10))
+
+    var authorId = 0;
+    if (auth && auth.currentUser && auth.currentUser.uid) {
+        authorId = auth.currentUser.uid;
+    }
+
+    // build query
+    var postsQuery
+    var conditions = [orderBy("createTime", "desc"), where("del", "==", trash.active), limit(10)]
+    if (searchKey) {
+        conditions.push(where("title", "==", searchKey))
+    }
+    if (trash.active) {
+        conditions.push(where("authorId", "==", authorId))
+    }
+    postsQuery = query(collectionRef, ...conditions);
 
     const [delIds, setDelIds] = useState([])
+    const [softDelIds, setSoftDelIds] = useState([])
+    const [restoringIds, setRestoringIds] = useState([])
+
+    const [isInTrash, setIsInTrash] = useState(false)
 
     const {isLoading, isFetching, error, data, fetchNextPage, refetch,  } = useFirestoreInfiniteQuery("posts", postsQuery, (snapshot) => {
         const lastDocument = snapshot.docs[snapshot.docs.length - 1];
@@ -39,13 +65,60 @@ const ListPost = ({isAuth}) => {
 
       useEffect(() => {
         document.getElementById('allPost')?.parentElement?.scrollTo(0,scroll.position.posY);
-      },)
-      
+      }, [])
 
-    const deletePost = async (id) => {
-        const postDoc = doc(db, "posts", id)
-        setDelIds([...delIds, id])
-        await deleteDoc(postDoc)
+      useEffect(() => {
+        setSearchKey(search.searchKey)
+      }, [search.searchKey])
+
+      const clearAll = () => {
+        setDelIds([])
+        setSoftDelIds([])
+        setRestoringIds([])
+      }
+
+      // on searching...
+      useEffect(() => {
+        doSearch()
+      }, [searchKey])
+
+      const doSearch = () => {
+        clearAll()
+        refetch()
+      }
+
+      // on moving to trash...
+      useEffect(() => {
+        movingToTrashNow(trash.active)
+      }, [trash.active])
+      
+      const movingToTrashNow = async (isInTrash)  => {
+        clearAll()
+        setIsInTrash(!isInTrash)
+        await refetch()
+        setIsInTrash(isInTrash)
+      }
+
+    const deletePost = async (id, del) => {
+        if (del) {
+            // permanent delete !
+            const postDoc = doc(db, "posts", id)
+            setDelIds([...delIds, id])
+            await deleteDoc(postDoc)
+        } else {
+            // soft del !
+            const updateDelDocRef = doc(db, 'posts', id);
+            setSoftDelIds([...softDelIds, id])
+            await updateDoc(updateDelDocRef, { del: true })
+        }
+        
+        refetch()
+    }
+
+    const restorePost = async (id) => {
+        const updateDelDocRef = doc(db, 'posts', id);
+        setRestoringIds([...restoringIds, id])
+        await updateDoc(updateDelDocRef, { del: false })
         refetch()
     }
 
@@ -77,15 +150,37 @@ const ListPost = ({isAuth}) => {
                             </Text>
                             <Flex mt='4'>
                                 <Spacer />
-                                {isAuth && post.author.id === auth.currentUser.uid && delIds.indexOf(post.id) === -1
+                                {isAuth && post.author.id === auth.currentUser.uid
+                                && delIds.indexOf(post.id) === -1
+                                && softDelIds.indexOf(post.id) === -1
+                                && restoringIds.indexOf(post.id) === -1
+                                && isInTrash
+                                    &&
+                                    (<>
+                                        <Button mr='2' as='button' borderRadius='md' bg='skyblue' color='white' px={4} h={8} onClick={() => { restorePost(post.id) }}> <ArrowUpIcon  /> </Button>
+                                    </>)
+                                }
+
+                                {isAuth && post.author.id === auth.currentUser.uid
+                                && delIds.indexOf(post.id) === -1
+                                && softDelIds.indexOf(post.id) === -1
+                                && restoringIds.indexOf(post.id) === -1
                                     &&
                                     (<>
                                         <Button mr='2' as='button' borderRadius='md' bg='yellowgreen' color='white' px={4} h={8} onClick={() => { editPost(post.id) }}> <EditIcon  /> </Button>
-                                        <Button as='button' borderRadius='md' bg='tomato' color='white' px={4} h={8} onClick={() => { deletePost(post.id) }}> <CloseIcon  /> </Button>
+                                        <Button as='button' borderRadius='md' bg='tomato' color='white' px={4} h={8} onClick={() => { deletePost(post.id, post.del) }}> 
+                                            {isInTrash ? <CloseIcon/> : <DeleteIcon/> }
+                                         </Button>
                                     </>)
                                     }
                                 {
-                                    (delIds.indexOf(post.id) !== -1) && <Center>Deleting...</Center>
+                                    (isInTrash && delIds.indexOf(post.id) !== -1) && <Center>Deleting...</Center>
+                                }
+                                {
+                                    (isInTrash && restoringIds.indexOf(post.id) !== -1) && <Center>Restoring...</Center>
+                                }
+                                {
+                                    (!isInTrash && softDelIds.indexOf(post.id) !== -1) && <Center>Moving to trash...</Center>
                                 }
                             </Flex>
                         </Box>
@@ -107,6 +202,16 @@ const ListPost = ({isAuth}) => {
 
 function Home({isAuth, setIsAuth}) {
     const navigate = useNavigate()
+    const dispatch = useDispatch()
+
+    const isInTrash = useSelector((state) => state.trash.value);
+    const search = useSelector((state) => state.search.value);
+    const [searchKey, setSearchKey] = useState(search.searchKey)
+
+    useEffect(() => {
+        dispatch(setKey(searchKey))
+    }, [searchKey])
+    
 
     const loginButton = () => {
     navigate('/kuteblog/login')
@@ -124,14 +229,21 @@ function Home({isAuth, setIsAuth}) {
         })
     }
 
+    const toggleTrash = () => {
+        dispatch(openTrash(!isInTrash.active))
+    }
+
   return (
     <ABasePage showBack={false} 
                 nav = {<>
                     {!isAuth && <Button onClick={loginButton}>Login</Button>}
                     {isAuth && <Button onClick={logout}>Log out</Button>}
                     <Spacer/>
-                    {isAuth && <Button onClick={() => {navigate('/kuteblog/trash')}} bg='tomato' color={'white'} mr='2'>Trash</Button>}
-                    {isAuth && <Button onClick={createPostButton} bg='yellowgreen' color={'white'}>Create Post</Button>}
+                    {isAuth && <Input ml='2' mr='2' value={searchKey} placeholder='Search...' onChange={(event) => {setSearchKey(event.target.value)}} ></Input> }
+                    {isAuth && <Button onClick={() => { toggleTrash() }} bg={isInTrash.active ? 'skyblue' : 'tomato'} color={'white'} mr='2'>
+                        {isInTrash.active ? <ArrowUpIcon/> : <DeleteIcon/> }
+                    </Button>}
+                    {isAuth && <Button onClick={createPostButton} bg='yellowgreen' color={'white'}> <EditIcon/> </Button>}
                 </>}
                 content={<ListPost isAuth={isAuth} />}
                >
